requirements.txt
Django==4.1.7
gunicorn==20.1.0
openai~=0.27.0
python-dotenv~=1.0.0
dj-database-url~=1.2.0
djangorestframework~=3.14.0
tiktoken~=0.3.0
djangorestframework-simplejwt~=5.2.2
mysqlclient~=2.1.1
django-allauth~=0.53.0
dj-rest-auth~=3.0.0
tls-client~=0.1.8
requests~=2.28.2
colorama~=0.4.4
svglib~=1.5.1
bs4~=0.0.1
beautifulsoup4~=4.11.2
reportlab~=3.6.12